+++ include/JSystem/JKernel/JKRDvdAramRipper.h
@@ -34,7 +34,7 @@ public:
 class JKRDvdFile;
 class JKRDvdAramRipper {
 public:
-    static JKRAramBlock* loadToAram(long, u32, JKRExpandSwitch, u32, u32);
+    static JKRAramBlock* loadToAram(s32, u32, JKRExpandSwitch, u32, u32);
     static JKRAramBlock* loadToAram(JKRDvdFile*, u32, JKRExpandSwitch, u32, u32);
     static JKRADCommand* loadToAram_Async(JKRDvdFile*, u32, JKRExpandSwitch, void (*)(u32), u32,
                                           u32);
