+++ include/JSystem/JKernel/JKRMemArchive.h
@@ -11,13 +11,13 @@ enum JKRMemBreakFlag {
 class JKRMemArchive : public JKRArchive {
 public:
     JKRMemArchive();
-    JKRMemArchive(long, JKRArchive::EMountDirection);
+    JKRMemArchive(s32, JKRArchive::EMountDirection);
     virtual ~JKRMemArchive();
 
     void fixedInit(s32);
     bool mountFixed(void*, JKRMemBreakFlag);
     void unmountFixed();
-    bool open(long, JKRArchive::EMountDirection);
+    bool open(s32, JKRArchive::EMountDirection);
     bool open(void*, u32, JKRMemBreakFlag);
 
     /* vt[09] */ void removeResourceAll(void); /* override */
