+++ src/JSystem/JKernel/JKRFile.cpp
@@ -8,7 +8,7 @@
 #include "dolphin/vi/vi.h"
 
 /* 802BC610-802BC6B4       .text read__7JKRFileFPvll */
-s32 JKRFile::read(void* data, s32 length, long offset) {
+s32 JKRFile::read(void* data, s32 length, s32 offset) {
     JUT_ASSERT(34, ( length & 0x1f ) == 0);
     while (true) {
         s32 result = readData(data, length, offset);
